+++ XamarinSDK/CobrowseIO.Xamarin.Abstractions/ICobrowseIO.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 
 namespace Xamarin.CobrowseIO.Abstractions
 {
@@ -34,13 +35,25 @@ namespace Xamarin.CobrowseIO.Abstractions
         void CreateSession(CobrowseCallback callback);
 
         /// <summary>
+        /// Gets or sets the API string.
+        /// </summary>
+        string Api { get; set; }
+
+        /// <summary>
         /// Gets the current Cobrowse.io device ID.
         /// </summary>
         string DeviceId { get; }
 
         /// <summary>
+        /// Gets or sets the license.
+        /// </summary>
+        string License { get; set; }
+
+        /// <summary>
         /// Sets the license.
         /// </summary>
+        [Obsolete("Use License property instead")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         void SetLicense(string licenseKey);
 
         /// <summary>
@@ -54,8 +67,15 @@ namespace Xamarin.CobrowseIO.Abstractions
         void Stop();
 
         /// <summary>
+        /// Gets or sets Cobrowse.io custom data. 
+        /// </summary>
+        IReadOnlyDictionary<string, object> CustomData { get; set; }
+
+        /// <summary>
         /// Sets Cobrowse.io custom data.
         /// </summary>
+        [Obsolete("Use CustomData property instead")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         void SetCustomData(IDictionary<string, object> customData);
 
         /// <summary>
@@ -66,6 +86,8 @@ namespace Xamarin.CobrowseIO.Abstractions
         /// <summary>
         /// Checks if full-device screen sharing is allowed.
         /// </summary>
-        void CheckCobrowseFullDevice();
+        [Obsolete("Use 'CobrowseAccessibilityService' directly in the Android project")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        bool CheckCobrowseFullDevice();
     }
 }