+++ iOS/CobrowseIO.iOS/CobrowseIO.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.ComponentModel;
 using Foundation;
 
 namespace Xamarin.CobrowseIO
@@ -20,11 +22,38 @@ namespace Xamarin.CobrowseIO
 
         public static CobrowseIO Instance => GetInstance();
 
+        public IReadOnlyDictionary<string, object> CustomData
+        {
+            get
+            {
+                if (this.CustomNSDictionaryData is NSDictionary dictionary)
+                {
+                    var rvalue = new Dictionary<string, object>();
+                    foreach (KeyValuePair<NSObject, NSObject> next in dictionary)
+                    {
+                        rvalue.Add(next.Key.ToString(), next.Value);
+                    }
+                    return rvalue;
+                }
+                return null;
+            }
+            set => SetCustomData(value);
+        }
+
+        [Obsolete("Use CustomData property instead")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public void SetCustomData(IDictionary<string, string> customData)
+        {
+            this.SetCustomData(
+                (IReadOnlyDictionary<string, string>)
+                new ReadOnlyDictionary<string, string>(customData));
+        }
+
+        internal void SetCustomData(IReadOnlyDictionary<string, string> customData)
         {
             if (customData == null)
             {
-                this.CustomData = null;
+                this.CustomNSDictionaryData = null;
                 return;
             }
             NSString[] objects = new NSString[customData.Count];
@@ -36,19 +65,28 @@ namespace Xamarin.CobrowseIO
                 objects[counter] = new NSString(next.Value);
                 counter++;
             }
-            this.CustomData = NSDictionary<NSString, NSObject>.FromObjectsAndKeys(objects, keys, customData.Count);
+            this.CustomNSDictionaryData = NSDictionary<NSString, NSObject>.FromObjectsAndKeys(objects, keys, customData.Count);
         }
 
         public void SetCustomData(NSDictionary<NSString, NSObject> customData)
         {
-            this.CustomData = customData;
+            this.CustomNSDictionaryData = customData;
         }
 
+        [Obsolete("Use CustomData property instead")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public void SetCustomData(IDictionary<string, object> customData)
+        {
+            this.SetCustomData(
+                (IReadOnlyDictionary<string, object>)
+                new ReadOnlyDictionary<string, object>(customData));
+        }
+
+        internal void SetCustomData(IReadOnlyDictionary<string, object> customData)
         {
             if (customData == null)
             {
-                this.CustomData = null;
+                this.CustomNSDictionaryData = null;
                 return;
             }
             NSObject[] objects = new NSObject[customData.Count];
@@ -84,10 +122,18 @@ namespace Xamarin.CobrowseIO
 
                 counter++;
             }
-            this.CustomData = NSDictionary<NSString, NSObject>.FromObjectsAndKeys(objects, keys, customData.Count);
+            this.CustomNSDictionaryData = NSDictionary<NSString, NSObject>.FromObjectsAndKeys(objects, keys, customData.Count);
         }
 
+        [Obsolete("Use Api property instead")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
+        public void SetApi(string api)
+        {
+            this.Api = api;
+        }
 
+        [Obsolete("Use License property instead")]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public void SetLicense(string license)
         {
             this.License = license;