+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [3.1.0]
+- Add new `CobrowseIO.License`, `CobrowseIO.CustomData`, `CobrowseIO.Api` properties
+- Add new `Session.Agent` property to the cross-platform wrapper API
+- The SDK now has `LinkerSafe` attribute
+- Android Native SDK [`2.5.0`](https://github.com/cobrowseio/cobrowse-sdk-android-binary/blob/master/CHANGELOG.md#250---2020-08-24)
+- iOS Native SDK [`2.8.1`](https://github.com/cobrowseio/cobrowse-sdk-ios-binary/blob/master/CHANGELOG.md#280---2020-08-24)
+
 ## [3.0.0] - 2020-05-07
 - Add cross-platform wrapper API (.NET Standard 1.0)
 - Android Native SDK [`2.2.0`](https://github.com/cobrowseio/cobrowse-sdk-android-binary/blob/master/CHANGELOG.md#200---2019-11-04)